Move Serilog related logging classes into Logging.Serilog
