+++ Assets/Exanite.Arpg/Logging/Serilog.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e4d37932c91717549bf1fcd69a68bbae
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
