+++ Assets/Exanite.Arpg/Logging/Serilog/SerilogLogAdapter.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4521a60dbd0c6894c8a73a7ef7abf93e
+guid: 6448945532b85244fbc004c7745badda
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
