+++ Assets/Exanite.Arpg/Logging/Serilog/ShortContextEnricher.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: eb571189bf5152a49ae2e8d5a7bcc2fc
+guid: f354aafa1fa43ec4f917bb6c3e0964cc
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
