+++ Assets/Exanite.Arpg/Logging/Serilog/UnityConsoleSink.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7b260900fa78ed04d8598295b0d7b8b3
+guid: 08f7706827d1f1241994e21a950b89a1
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
