+++ Assets/Exanite.Arpg/Logging/Serilog/UnityToSerilogLogHandler.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 33c8b481c3de83c45a406a7fe4e894f6
+guid: 8c1e77caf7e627147a10e8bc057f4623
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
